+++ src/common/config.h
@@ -15,7 +15,7 @@ namespace marian {
 class Config {
 public:
   static size_t seed;
-
+  
   typedef YAML::Node YamlNode;
 
   Config(const std::string options,
@@ -45,6 +45,8 @@ public:
                   bool validate = true) {
     auto parser = ConfigParser(argc, argv, mode, validate);
     config_ = parser.getConfig();
+    devices_ = parser.getDevices();
+    
     createLoggers(this);
 
     if(get<size_t>("seed") == 0)
@@ -100,6 +102,10 @@ public:
   YAML::Node getModelParameters();
   void loadModelParameters(const std::string& name);
 
+  const std::vector<DeviceId>& getDevices() {
+    return devices_;
+  }
+  
   void save(const std::string& name) {
     OutputFileStream out(name);
     (std::ostream&)out << *this;
@@ -118,6 +124,7 @@ public:
 
 private:
   YAML::Node config_;
+  std::vector<DeviceId> devices_;
 
   static void GetYamlFromNpz(YAML::Node&,
                              const std::string&,