+++ src/common/config_parser.cpp
@@ -5,7 +5,19 @@
 #include <regex>
 #include <stdexcept>
 
+#if MKL_FOUND
+//#include <omp.h>
+#include <mkl.h>
+#else
+#if BLAS_FOUND
+//#include <omp.h>
+#include <cblas.h>
+#endif
+#endif
+
+
 #include "3rd_party/cnpy/cnpy.h"
+#include "common/definitions.h"
 #include "common/config.h"
 #include "common/config_parser.h"
 #include "common/file_stream.h"
@@ -428,8 +440,10 @@ void ConfigParser::addOptionsTraining(po::options_description& desc) {
       ->multitoken()
       ->default_value(std::vector<std::string>({"0"}), "0"),
       "GPU ID(s) to use for training")
-    ("cpu", po::value<bool>()->zero_tokens()->default_value(false),
-      "Use CPU-based training or decoding")
+    ("cpu-threads", po::value<size_t>()->default_value(0)->implicit_value(1),
+      "Use CPU-based computation with this many independent threads, 0 means GPU-based computation")
+    //("omp-threads", po::value<size_t>()->default_value(1),
+    //  "Set number of OpenMP threads for each CPU-based thread")
 
     ("mini-batch", po::value<int>()->default_value(64),
       "Size of mini-batch used during update")
@@ -438,6 +452,8 @@ void ConfigParser::addOptionsTraining(po::options_description& desc) {
     ("mini-batch-fit", po::value<bool>()->zero_tokens()->default_value(false),
       "Determine mini-batch size automatically based on sentence-length to "
       "fit reserved memory")
+    ("mini-batch-fit-step", po::value<size_t>()->default_value(10),
+      "Step size for mini-batch-fit statistics")
     ("maxi-batch", po::value<int>()->default_value(100),
       "Number of batches to preload for length-based sorting")
     ("maxi-batch-sort", po::value<std::string>()->default_value("trg"),
@@ -616,8 +632,11 @@ void ConfigParser::addOptionsTranslate(po::options_description& desc) {
       ->multitoken()
       ->default_value(std::vector<std::string>({"0"}), "0"),
       "GPUs to use for translating")
-    ("cpu", po::value<bool>()->zero_tokens()->default_value(false),
-      "Use CPU-based decoding")
+    ("cpu-threads", po::value<size_t>()->default_value(0)->implicit_value(1),
+      "Use CPU-based computation with this many independent threads, 0 means GPU-based computation")
+    //("omp-threads", po::value<size_t>()->default_value(1),
+    //  "Set number of OpenMP threads for each CPU-based thread")
+
     ("mini-batch", po::value<int>()->default_value(1),
       "Size of mini-batch used during update")
     ("maxi-batch", po::value<int>()->default_value(1),
@@ -662,17 +681,19 @@ void ConfigParser::addOptionsRescore(po::options_description& desc) {
       ->multitoken()
       ->default_value(std::vector<std::string>({"0"}), "0"),
       "GPUs to use for training")
-    ("cpu", po::value<bool>()->zero_tokens()->default_value(false),
-      "Use CPU-based scoring")
+    ("cpu-threads", po::value<size_t>()->default_value(0)->implicit_value(1),
+      "Use CPU-based computation with this many independent threads, 0 means GPU-based computation")
+    //("omp-threads", po::value<size_t>()->default_value(1),
+    //  "Set number of OpenMP threads for each CPU-based thread")
 
     ("mini-batch", po::value<int>()->default_value(64),
       "Size of mini-batch used during update")
     ("mini-batch-words", po::value<int>()->default_value(0),
       "Set mini-batch size based on words instead of sentences")
-    ("mini-batch-fit", po::value<bool>()->zero_tokens()->default_value(false),
-      "Determine mini-batch size automatically based on sentence-length to fit reserved memory")
     ("maxi-batch", po::value<int>()->default_value(100),
       "Number of batches to preload for length-based sorting")
+    ("maxi-batch-sort", po::value<std::string>()->default_value("trg"),
+      "Sorting strategy for maxi-batch: trg (default) src none")
     ;
   // clang-format on
   desc.add(rescore);
@@ -840,6 +861,7 @@ void ConfigParser::parseOptions(int argc, char** argv, bool doValidate) {
     SET_OPTION("sync-sgd", bool);
     SET_OPTION("mini-batch-words", int);
     SET_OPTION("mini-batch-fit", bool);
+    SET_OPTION("mini-batch-fit-step", size_t);
 
     SET_OPTION("lr-decay", double);
     SET_OPTION("lr-decay-strategy", std::string);
@@ -885,7 +907,6 @@ void ConfigParser::parseOptions(int argc, char** argv, bool doValidate) {
       config_["train-sets"] = vm_["train-sets"].as<std::vector<std::string>>();
     }
     SET_OPTION("mini-batch-words", int);
-    SET_OPTION("mini-batch-fit", bool);
     SET_OPTION_NONDEFAULT("summary", std::string);
   }
 
@@ -930,13 +951,13 @@ void ConfigParser::parseOptions(int argc, char** argv, bool doValidate) {
   SET_OPTION("seed", size_t);
   SET_OPTION("relative-paths", bool);
   SET_OPTION("devices", std::vector<std::string>);
-  SET_OPTION("cpu", bool);
+  SET_OPTION("cpu-threads", size_t);
+  //SET_OPTION("omp-threads", size_t);
 
   SET_OPTION("mini-batch", int);
   SET_OPTION("maxi-batch", int);
 
-  if(mode_ == ConfigMode::training || mode_ == ConfigMode::translating)
-    SET_OPTION("maxi-batch-sort", std::string);
+  SET_OPTION("maxi-batch-sort", std::string);
   SET_OPTION("max-length", size_t);
   SET_OPTION("max-length-crop", bool);
 
@@ -976,36 +997,55 @@ void ConfigParser::parseOptions(int argc, char** argv, bool doValidate) {
     exit(0);
   }
 
-  try {
-    processOptionDevices();
-  } catch(const std::invalid_argument& e) {
-    ABORT("Conversion of --devices option failed, please report a bug");
-  }
+// @TODO: this should probably be in processOptionDevices()
+//#ifdef BLAS_FOUND
+//  //omp_set_num_threads(vm_["omp-threads"].as<size_t>());
+//#ifdef MKL_FOUND
+//  mkl_set_num_threads(vm_["omp-threads"].as<size_t>());
+//#endif
+//#endif
 }
 
-void ConfigParser::processOptionDevices() {
-  std::string devicesStr
-      = Join(config_["devices"].as<std::vector<std::string>>());
-  std::vector<size_t> devices;
+std::vector<DeviceId> ConfigParser::getDevices() {
+  std::vector<DeviceId> devices;
 
-  if(mode_ == ConfigMode::training && get<bool>("multi-node")) {
-    auto parts = Split(devicesStr, ":");
-    for(size_t i = 1; i < parts.size(); ++i) {
-      std::string part = parts[i];
-      Trim(part);
-      auto ds = Split(part, " ");
-      if(i < parts.size() - 1)
-        ds.pop_back();
-      devices.emplace_back(ds.size());
-      for(auto d : ds)
-        devices.emplace_back(std::stoi(d));
+  try {
+    
+    std::string devicesStr
+        = Join(config_["devices"].as<std::vector<std::string>>());
+        
+    
+    if(mode_ == ConfigMode::training && get<bool>("multi-node")) {
+      auto parts = Split(devicesStr, ":");
+      for(size_t i = 1; i < parts.size(); ++i) {
+        std::string part = parts[i];
+        Trim(part);
+        auto ds = Split(part, " ");
+        if(i < parts.size() - 1)
+          ds.pop_back();
+        
+        // does this make sense?
+        devices.push_back({ds.size(), DeviceType::gpu});
+        for(auto d : ds)
+          devices.push_back({std::stoull(d), DeviceType::gpu});
+      }
+    } else {
+      for(auto d : Split(devicesStr))
+        devices.push_back({std::stoull(d), DeviceType::gpu});
     }
-  } else {
-    for(auto d : Split(devicesStr))
-      devices.emplace_back(std::stoi(d));
+    
+    if(config_["cpu-threads"].as<size_t>() > 0) {
+      devices.clear();
+      for(size_t i = 0; i < config_["cpu-threads"].as<size_t>(); ++i)
+      devices.push_back({i, DeviceType::cpu});
+    }
+    
+  }
+  catch(...) {
+    ABORT("Problem parsing devices, please report an issue on github");
   }
 
-  config_["devices"] = devices;
+  return devices;
 }
 
 YAML::Node ConfigParser::getConfig() const {