+++ src/graph/node_operators_binary.h
@@ -2,11 +2,10 @@
 
 #include <thread>
 
-#include "tensors/gpu/backend.h"
 #include "graph/node.h"
 #include "functional/functional.h"
-#include "kernels/tensor_operators.h"
-#include "kernels/cudnn_wrappers.h"
+#include "tensors/tensor_operators.h"
+#include "tensors/gpu/cudnn_wrappers.h"
 
 namespace marian {
 
@@ -54,7 +53,6 @@ public:
   NodeOps forwardOps() {
     // C = alpha * dot(op(A), op(B))
     return {NodeOp(Prod(
-        std::static_pointer_cast<gpu::Backend>(getBackend())->getCublasHandle(),
         val_,
         child(0)->val(),
         child(1)->val(),
@@ -72,18 +70,14 @@ public:
     // to sum gradients from different graph parts
 
     if(!transA_ && transB_)
-      return {NodeOp(Prod(std::static_pointer_cast<gpu::Backend>(getBackend())
-                              ->getCublasHandle(),
-                          child(0)->grad(),
+      return {NodeOp(Prod(child(0)->grad(),
                           adj_,
                           child(1)->val(),
                           false,
                           false,
                           1.0,
                           scalar_)),
-              NodeOp(Prod(std::static_pointer_cast<gpu::Backend>(getBackend())
-                              ->getCublasHandle(),
-                          child(1)->grad(),
+              NodeOp(Prod(child(1)->grad(),
                           adj_,
                           child(0)->val(),
                           true,
@@ -92,18 +86,14 @@ public:
                           scalar_))};
 
     if(transA_ && !transB_)
-      return {NodeOp(Prod(std::static_pointer_cast<gpu::Backend>(getBackend())
-                              ->getCublasHandle(),
-                          child(0)->grad(),
+      return {NodeOp(Prod(child(0)->grad(),
                           child(1)->val(),
                           adj_,
                           false,
                           true,
                           1.0,
                           scalar_)),
-              NodeOp(Prod(std::static_pointer_cast<gpu::Backend>(getBackend())
-                              ->getCublasHandle(),
-                          child(1)->grad(),
+              NodeOp(Prod(child(1)->grad(),
                           child(0)->val(),
                           adj_,
                           false,
@@ -112,18 +102,14 @@ public:
                           scalar_))};
 
     if(transA_ && transB_)
-      return {NodeOp(Prod(std::static_pointer_cast<gpu::Backend>(getBackend())
-                              ->getCublasHandle(),
-                          child(0)->grad(),
+      return {NodeOp(Prod(child(0)->grad(),
                           child(1)->val(),
                           adj_,
                           true,
                           true,
                           1.0,
                           scalar_)),
-              NodeOp(Prod(std::static_pointer_cast<gpu::Backend>(getBackend())
-                              ->getCublasHandle(),
-                          child(1)->grad(),
+              NodeOp(Prod(child(1)->grad(),
                           adj_,
                           child(0)->val(),
                           true,
@@ -131,18 +117,14 @@ public:
                           1.0,
                           scalar_))};
 
-    return {NodeOp(Prod(std::static_pointer_cast<gpu::Backend>(getBackend())
-                            ->getCublasHandle(),
-                        child(0)->grad(),
+    return {NodeOp(Prod(child(0)->grad(),
                         adj_,
                         child(1)->val(),
                         false,
                         true,
                         1.0,
                         scalar_)),
-            NodeOp(Prod(std::static_pointer_cast<gpu::Backend>(getBackend())
-                            ->getCublasHandle(),
-                        child(1)->grad(),
+            NodeOp(Prod(child(1)->grad(),
                         child(0)->val(),
                         adj_,
                         true,
@@ -198,7 +180,6 @@ public:
     using namespace functional;
     return {
       NodeOp(Prod(
-        std::static_pointer_cast<gpu::Backend>(getBackend())->getCublasHandle(),
         val_,
         child(0)->val(),
         child(1)->val(),
@@ -219,18 +200,14 @@ public:
     using namespace functional;
 
     if(!transA_ && transB_)
-      return {NodeOp(Prod(std::static_pointer_cast<gpu::Backend>(getBackend())
-                              ->getCublasHandle(),
-                          child(0)->grad(),
+      return {NodeOp(Prod(child(0)->grad(),
                           adj_,
                           child(1)->val(),
                           false,
                           false,
                           1.0,
                           scalar_)),
-              NodeOp(Prod(std::static_pointer_cast<gpu::Backend>(getBackend())
-                              ->getCublasHandle(),
-                          child(1)->grad(),
+              NodeOp(Prod(child(1)->grad(),
                           adj_,
                           child(0)->val(),
                           true,
@@ -240,18 +217,14 @@ public:
               NodeOp(Add(_1, child(2)->grad(), adj_))};
 
     if(transA_ && !transB_)
-      return {NodeOp(Prod(std::static_pointer_cast<gpu::Backend>(getBackend())
-                              ->getCublasHandle(),
-                          child(0)->grad(),
+      return {NodeOp(Prod(child(0)->grad(),
                           child(1)->val(),
                           adj_,
                           false,
                           true,
                           1.0,
                           scalar_)),
-              NodeOp(Prod(std::static_pointer_cast<gpu::Backend>(getBackend())
-                              ->getCublasHandle(),
-                          child(1)->grad(),
+              NodeOp(Prod(child(1)->grad(),
                           child(0)->val(),
                           adj_,
                           false,
@@ -261,18 +234,14 @@ public:
               NodeOp(Add(_1, child(2)->grad(), adj_))};
 
     if(transA_ && transB_)
-      return {NodeOp(Prod(std::static_pointer_cast<gpu::Backend>(getBackend())
-                              ->getCublasHandle(),
-                          child(0)->grad(),
+      return {NodeOp(Prod(child(0)->grad(),
                           child(1)->val(),
                           adj_,
                           true,
                           true,
                           1.0,
                           scalar_)),
-              NodeOp(Prod(std::static_pointer_cast<gpu::Backend>(getBackend())
-                              ->getCublasHandle(),
-                          child(1)->grad(),
+              NodeOp(Prod(child(1)->grad(),
                           adj_,
                           child(0)->val(),
                           true,
@@ -281,18 +250,14 @@ public:
                           scalar_)),
               NodeOp(Add(_1, child(2)->grad(), adj_))};
 
-    return {NodeOp(Prod(std::static_pointer_cast<gpu::Backend>(getBackend())
-                            ->getCublasHandle(),
-                        child(0)->grad(),
+    return {NodeOp(Prod(child(0)->grad(),
                         adj_,
                         child(1)->val(),
                         false,
                         true,
                         1.0,
                         scalar_)),
-            NodeOp(Prod(std::static_pointer_cast<gpu::Backend>(getBackend())
-                            ->getCublasHandle(),
-                        child(1)->grad(),
+            NodeOp(Prod(child(1)->grad(),
                         child(0)->val(),
                         adj_,
                         true,
@@ -350,7 +315,6 @@ public:
   NodeOps forwardOps() {
     // C = alpha * dot(op(A), op(B))
     return {NodeOp(ProdBatched(
-        std::static_pointer_cast<gpu::Backend>(getBackend())->getCublasHandle(),
         val_,
         child(0)->val(),
         child(1)->val(),
@@ -369,18 +333,14 @@ public:
 
     if(!transA_ && transB_)
       return {
-          NodeOp(ProdBatched(std::static_pointer_cast<gpu::Backend>(getBackend())
-                                 ->getCublasHandle(),
-                             child(0)->grad(),
+          NodeOp(ProdBatched(child(0)->grad(),
                              adj_,
                              child(1)->val(),
                              false,
                              false,
                              1.0,
                              scalar_)),
-          NodeOp(ProdBatched(std::static_pointer_cast<gpu::Backend>(getBackend())
-                                 ->getCublasHandle(),
-                             child(1)->grad(),
+          NodeOp(ProdBatched(child(1)->grad(),
                              adj_,
                              child(0)->val(),
                              true,
@@ -390,18 +350,14 @@ public:
 
     if(transA_ && !transB_)
       return {
-          NodeOp(ProdBatched(std::static_pointer_cast<gpu::Backend>(getBackend())
-                                 ->getCublasHandle(),
-                             child(0)->grad(),
+          NodeOp(ProdBatched(child(0)->grad(),
                              child(1)->val(),
                              adj_,
                              false,
                              true,
                              1.0,
                              scalar_)),
-          NodeOp(ProdBatched(std::static_pointer_cast<gpu::Backend>(getBackend())
-                                 ->getCublasHandle(),
-                             child(1)->grad(),
+          NodeOp(ProdBatched(child(1)->grad(),
                              child(0)->val(),
                              adj_,
                              false,
@@ -411,18 +367,14 @@ public:
 
     if(transA_ && transB_)
       return {
-          NodeOp(ProdBatched(std::static_pointer_cast<gpu::Backend>(getBackend())
-                                 ->getCublasHandle(),
-                             child(0)->grad(),
+          NodeOp(ProdBatched(child(0)->grad(),
                              child(1)->val(),
                              adj_,
                              true,
                              true,
                              1.0,
                              scalar_)),
-          NodeOp(ProdBatched(std::static_pointer_cast<gpu::Backend>(getBackend())
-                                 ->getCublasHandle(),
-                             child(1)->grad(),
+          NodeOp(ProdBatched(child(1)->grad(),
                              adj_,
                              child(0)->val(),
                              true,
@@ -431,18 +383,14 @@ public:
                              scalar_))};
 
     return {
-        NodeOp(ProdBatched(std::static_pointer_cast<gpu::Backend>(getBackend())
-                               ->getCublasHandle(),
-                           child(0)->grad(),
+        NodeOp(ProdBatched(child(0)->grad(),
                            adj_,
                            child(1)->val(),
                            false,
                            true,
                            1.0,
                            scalar_)),
-        NodeOp(ProdBatched(std::static_pointer_cast<gpu::Backend>(getBackend())
-                               ->getCublasHandle(),
-                           child(1)->grad(),
+        NodeOp(ProdBatched(child(1)->grad(),
                            child(0)->val(),
                            adj_,
                            true,