+++ src/optimizers/optimizers.cu
@@ -1,7 +1,6 @@
 #include "optimizers.h"
 
-#include "functional/functional.h"
-#include "kernels/tensor_operators.h"
+#include "tensors/tensor_operators.h"
 
 namespace marian {
 
@@ -9,7 +8,7 @@ void Sgd::updateImpl(Tensor params, Tensor grads) {
   using namespace functional;
   Element(_1 -= (multiplyFactor_ * eta_) * _2, params, grads);
 
-  cudaStreamSynchronize(0);
+  //cudaStreamSynchronize(0);
 }
 
 // Aagrad
@@ -34,7 +33,7 @@ void Adagrad::updateImpl(Tensor params, Tensor grads) {
           gt_,
           grads);
 
-  cudaStreamSynchronize(0);
+  //cudaStreamSynchronize(0);
 }
 
 void Adagrad::load(const std::string& name,
@@ -120,7 +119,7 @@ void Adagrad::save(const std::string& name,
 void Adagrad::resetStats() {
   if(gt_)
     gt_->set(0);
-  cudaStreamSynchronize(0);
+  //cudaStreamSynchronize(0);
 }
 
 // Adam
@@ -154,7 +153,7 @@ void Adam::updateImpl(Tensor params, Tensor grads) {
           mt_,
           vt_);
 
-  cudaStreamSynchronize(0);
+  //cudaStreamSynchronize(0);
 }
 
 void Adam::load(const std::string& name,
@@ -258,7 +257,7 @@ void Adam::resetStats() {
   if(vt_)
     vt_->set(0);
 
-  cudaStreamSynchronize(0);
+  //cudaStreamSynchronize(0);
 }
 
 Ptr<OptimizerBase> Optimizer(Ptr<Config> options) {